Move Rbkit::Profiler to Rbkit::Server in its own file
